+++ resources/views/event/edit.blade.php
@@ -14,11 +14,12 @@
     <div class="row justify-content-center">
         <div class="col-md-8">
             <div class="card">
-                <div class="card-header">Add new Event</div>
+                <div class="card-header">Edit Event</div>
 
                 <div class="card-body">
-                    <form method="POST" action="{{ route('events.store') }}">
+                    <form method="POST" action="{{ route('events.update', $event) }}">
                         @csrf
+                        @method('PATCH')
                         {{--Name--}}
                         <div class="form-group row">
                             <label for="name" class="col-md-4 col-form-label text-md-right"> Name</label>
@@ -26,7 +27,7 @@
                             <div class="col-md-6">
                                 <input id="name" type="text"
                                        class="form-control{{ $errors->has('name') ? ' is-invalid' : '' }}" name="name"
-                                       value="{{ old('name') }}" required autofocus>
+                                       value="{{ old('name', $event->name) }}" required autofocus>
 
                                 @if ($errors->has('name'))
                                     <span class="invalid-feedback">
@@ -45,7 +46,7 @@
                                         name="eventType">
                                     <option value="">Choose an event type...</option>
                                     @foreach($eventTypes as $eventType)
-                                        <option value="{{ $eventType->id }}" {{ old('eventType') == $eventType->id ? 'selected' : '' }}>{{ $eventType->name }}</option>
+                                        <option value="{{ $eventType->id }}" {{ old('eventType', $event->event_type_id) == $eventType->id ? 'selected' : '' }}>{{ $eventType->name }}</option>
                                     @endforeach
                                 </select>
 
@@ -65,12 +66,12 @@
                             <div class="col-md-6">
                                 <div class="custom-control custom-control-inline custom-radio">
                                     <input type="radio" value="1" id="import_only1" name="import_only"
-                                           class="custom-control-input" {{ old('import_only') == 1 ? 'checked' : '' }}>
+                                           class="custom-control-input" {{ old('import_only', $event->import_only) == 1 ? 'checked' : '' }}>
                                     <label class="custom-control-label" for="import_only1">Yes</label>
                                 </div>
                                 <div class="custom-control custom-control-inline custom-radio">
                                     <input type="radio" value="0" id="import_only0" name="import_only"
-                                           class="custom-control-input" {{ old('import_only') == 0 ? 'checked' : '' }}>
+                                           class="custom-control-input" {{ old('import_only', $event->import_only) == 0 ? 'checked' : '' }}>
                                     <label class="custom-control-label" for="import_only0">No</label>
                                 </div>
 
@@ -90,12 +91,12 @@
                             <div class="col-md-6">
                                 <div class="custom-control custom-control-inline custom-radio">
                                     <input type="radio" value="1" id="uses_times1" name="uses_times"
-                                           class="custom-control-input" {{ old('uses_times') == 1 ? 'checked' : '' }}>
+                                           class="custom-control-input" {{ old('uses_times', $event->uses_times) == 1 ? 'checked' : '' }}>
                                     <label class="custom-control-label" for="uses_times1">Yes</label>
                                 </div>
                                 <div class="custom-control custom-control-inline custom-radio">
                                     <input type="radio" value="0" id="uses_times0" name="uses_times"
-                                           class="custom-control-input" {{ old('uses_times') == 0 ? 'checked' : '' }}>
+                                           class="custom-control-input" {{ old('uses_times', $event->uses_times) == 0 ? 'checked' : '' }}>
                                     <label class="custom-control-label" for="uses_times0">No</label>
                                 </div>
 
@@ -116,13 +117,13 @@
                                 <div class="custom-control custom-control-inline custom-radio">
                                     <input type="radio" value="1" id="multiple_bookings_allowed1"
                                            name="multiple_bookings_allowed"
-                                           class="custom-control-input" {{ old('multiple_bookings_allowed') == 1 ? 'checked' : '' }}>
+                                           class="custom-control-input" {{ old('multiple_bookings_allowed', $event->multiple_bookings_allowed) == 1 ? 'checked' : '' }}>
                                     <label class="custom-control-label" for="multiple_bookings_allowed1">Yes</label>
                                 </div>
                                 <div class="custom-control custom-control-inline custom-radio">
                                     <input type="radio" value="0" id="multiple_bookings_allowed0"
                                            name="multiple_bookings_allowed"
-                                           class="custom-control-input" {{ old('multiple_bookings_allowed') == 0 ? 'checked' : '' }}>
+                                           class="custom-control-input" {{ old('multiple_bookings_allowed', $event->multiple_bookings_allowed) == 0 ? 'checked' : '' }}>
                                     <label class="custom-control-label" for="multiple_bookings_allowed0">No</label>
                                 </div>
 
@@ -142,12 +143,12 @@
                             <div class="col-md-6">
                                 <div class="custom-control custom-control-inline custom-radio">
                                     <input type="radio" value="1" id="is_oceanic_event1" name="is_oceanic_event"
-                                           class="custom-control-input" {{ old('is_oceanic_event') == 1 ? 'checked' : '' }}>
+                                           class="custom-control-input" {{ old('is_oceanic_event', $event->is_oceanic_event) == 1 ? 'checked' : '' }}>
                                     <label class="custom-control-label" for="is_oceanic_event1">Yes</label>
                                 </div>
                                 <div class="custom-control custom-control-inline custom-radio">
                                     <input type="radio" value="0" id="is_oceanic_event0" name="is_oceanic_event"
-                                           class="custom-control-input" {{ old('is_oceanic_event') == 0 ? 'checked' : '' }}>
+                                           class="custom-control-input" {{ old('is_oceanic_event', $event->is_oceanic_event) == 0 ? 'checked' : '' }}>
                                     <label class="custom-control-label" for="is_oceanic_event0">No</label>
                                 </div>
 
@@ -168,7 +169,7 @@
                                         name="dep">
                                     <option value="">Choose an airport...</option>
                                     @foreach($airports as $airport)
-                                        <option value="{{ $airport->icao }}" {{ old('dep') == $airport->icao ? 'selected' : '' }}>{{ $airport->icao }}
+                                        <option value="{{ $airport->icao }}" {{ old('dep', $event->airportDep->icao) == $airport->icao ? 'selected' : '' }}>{{ $airport->icao }}
                                             [{{ $airport->name }} ({{ $airport->iata }})]
                                         </option>
                                     @endforeach
@@ -191,7 +192,7 @@
                                         name="arr">
                                     <option value="">Choose an airport...</option>
                                     @foreach($airports as $airport)
-                                        <option value="{{ $airport->icao }}" {{ old('arr') == $airport->icao ? 'selected' : '' }}>{{ $airport->icao }}
+                                        <option value="{{ $airport->icao }}" {{ old('arr', $event->airportArr->icao) == $airport->icao ? 'selected' : '' }}>{{ $airport->icao }}
                                             [{{ $airport->name }} ({{ $airport->iata }})]
                                         </option>
                                     @endforeach
@@ -213,7 +214,7 @@
                             <div class="col-md-6">
                                 <input type="text"
                                        class="form-control{{ $errors->has('date') ? ' is-invalid' : '' }} datepicker"
-                                       name="dateEvent" value="{{ old('dateEvent') }}" required>
+                                       name="dateEvent" value="{{ old('dateEvent', $event->startEvent->format('d-m-Y')) }}" required>
 
                                 @if ($errors->has('dateEvent'))
                                     <span class="invalid-feedback">
@@ -231,7 +232,7 @@
                             <div class="col-md-6">
                                 <input id="timeBeginEvent" type="time"
                                        class="form-control{{ $errors->has('timeEndEvent') ? ' is-invalid' : '' }}"
-                                       name="timeBeginEvent" value="{{ old('timeBeginEvent') }}" required>
+                                       name="timeBeginEvent" value="{{ old('timeBeginEvent', $event->startEvent->format('H:i')) }}" required>
 
                                 @if ($errors->has('name'))
                                     <span class="invalid-feedback">
@@ -249,7 +250,7 @@
                             <div class="col-md-6">
                                 <input id="timeEndEvent" type="time"
                                        class="form-control{{ $errors->has('timeEndEvent') ? ' is-invalid' : '' }}"
-                                       name="timeEndEvent" value="{{ old('timeEndEvent') }}" required>
+                                       name="timeEndEvent" value="{{ old('timeEndEvent', $event->endEvent->format('H:i')) }}" required>
 
                                 @if ($errors->has('timeEndEvent'))
                                     <span class="invalid-feedback">
@@ -267,7 +268,7 @@
                             <div class="col-md-6">
                                 <input type="text"
                                        class="form-control{{ $errors->has('date') ? ' is-invalid' : '' }} datepicker"
-                                       name="dateBeginBooking" value="{{ old('dateBeginBooking') }}" required>
+                                       name="dateBeginBooking" value="{{ old('dateBeginBooking', $event->startBooking->format('d-m-Y')) }}" required>
 
                                 @if ($errors->has('dateBeginBooking'))
                                     <span class="invalid-feedback">
@@ -285,7 +286,7 @@
                             <div class="col-md-6">
                                 <input id="timeBeginEvent" type="time"
                                        class="form-control{{ $errors->has('timeBeginBooking') ? ' is-invalid' : '' }}"
-                                       name="timeBeginBooking" value="{{ old('timeBeginBooking') }}" required>
+                                       name="timeBeginBooking" value="{{ old('timeBeginBooking', $event->startBooking->format('H:i')) }}" required>
 
                                 @if ($errors->has('name'))
                                     <span class="invalid-feedback">
@@ -303,7 +304,7 @@
                             <div class="col-md-6">
                                 <input type="text"
                                        class="form-control{{ $errors->has('date') ? ' is-invalid' : '' }} datepicker"
-                                       name="dateEndBooking" value="{{ old('dateEndBooking') }}" required>
+                                       name="dateEndBooking" value="{{ old('dateEndBooking', $event->endBooking->format('d-m-Y')) }}" required>
 
                                 @if ($errors->has('dateBeginBooking'))
                                     <span class="invalid-feedback">
@@ -321,7 +322,7 @@
                             <div class="col-md-6">
                                 <input id="timeEndBooking" type="time"
                                        class="form-control{{ $errors->has('timeEndBooking') ? ' is-invalid' : '' }}"
-                                       name="timeEndBooking" value="{{ old('timeEndBooking') }}" required>
+                                       name="timeEndBooking" value="{{ old('timeEndBooking', $event->endBooking->format('H:i')) }}" required>
 
                                 @if ($errors->has('timeEndBooking'))
                                     <span class="invalid-feedback">
@@ -339,7 +340,7 @@
                                 <input id="image" type="url"
                                        class="form-control{{ $errors->has('image_url') ? ' is-invalid' : '' }}"
                                        name="image_url"
-                                       value="{{ old('image_url') }}">
+                                       value="{{ old('image_url', $event->image_url) }}">
 
                                 @if ($errors->has('image_url'))
                                     <span class="invalid-feedback">
@@ -352,14 +353,14 @@
                         {{-- Description --}}
                         <div class="form-group row">
                             <textarea id="description" name="description"
-                                      rows="10">{!! old(html_entity_decode('description')) !!}</textarea>
+                                      rows="10">{!! html_entity_decode(old('description', $event->description)) !!}</textarea>
                         </div>
 
-                        {{--Add--}}
+                        {{--Edit--}}
                         <div class="form-group row mb-0">
                             <div class="col-md-6 offset-md-4">
                                 <button type="submit" class="btn btn-primary">
-                                    <i class="fa fa-plus"></i> Add
+                                    <i class="fa fa-plus"></i> Edit
                                 </button>
                             </div>
                         </div>